Rename functions.py to calculatons.py
