+++ ldndctools/misc/calculations.py
